+++ uppsrc/CodeEditor/Syntax.h
@@ -111,11 +111,11 @@ public:
 	virtual void            ScanSyntax(const wchar *ln, const wchar *e, int line, int tab_size);
 	virtual void            Serialize(Stream& s);
 	virtual void            IndentInsert(CodeEditor& editor, int chr, int count);
-	virtual bool            CheckBrackets(CodeEditor& e, int& bpos0, int& bpos); // TODO: Replace with generic mechanism
-	virtual void            CheckSyntaxRefresh(CodeEditor& e, int pos, const WString& text);
+	virtual bool            CheckBrackets(CodeEditor& e, int64& bpos0, int64& bpos); // TODO: Replace with generic mechanism
+	virtual void            CheckSyntaxRefresh(CodeEditor& e, int64 pos, const WString& text);
 	virtual bool            CanAssist() const;
 	virtual void            Highlight(const wchar *s, const wchar *end, HighlightOutput& hls,
-	                                  CodeEditor *editor, int line, int pos);
+	                                  CodeEditor *editor, int line, int64 pos);
 	virtual Vector<IfState> PickIfStack();
 	virtual void            ReformatComment(CodeEditor& e);
 	virtual ~EditorSyntax();
